requirements.txt
asgiref==3.7.2
Django==4.2.2
django-phonenumber-field==7.1.0
gunicorn==20.1.0
tzdata==2022.1
# psycopg2-binary==2.9.3
sqlparse==0.4.4
phonenumbers==8.13.14
python-dotenv==1.0.0
loguru
djangorestframework
